Extract parser schema DDL grammar
